refactor: restructure util and lib functions
