+++ frontend/components/search/wizard/modes/llm/panels/side/index.tsx
@@ -6,7 +6,7 @@ import { ArrowLeft, Loader2, SlidersHorizontal, Wand2 } from "lucide-react";
 
 import { useSearchWizard } from "@contexts/search-wizard";
 import { useIntakeSessions } from "@hooks/use-intake-sessions";
-import { getApiErrorMessage } from "@lib/api-errors";
+import { getApiErrorMessage } from "@utils/common";
 import type { LlmInputResponse } from "@services/intake-sessions";
 
 import { STYLES } from "../../styles";
