+++ frontend/utils/common/index.ts
@@ -1,4 +1,5 @@
 export { cn } from "./cn";
+export { getApiErrorMessage } from "./api-errors";
 export {
   type FormatMoneyOptions,
   formatFeet,
