--- frontend/utils/listings/index.ts
@@ -1,2 +1,0 @@
-export { listingLocationLine, listingTitle } from "./headline";
-export { mapsHref } from "./maps";
